backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
redis==5.0.1
rq==1.15.1
openai==1.3.7
google-cloud-vision==3.4.5
twilio==8.10.3
boto3==1.34.0
minio==7.2.0
pillow==10.1.0
python-magic==0.4.27
slowapi==0.1.9
sentry-sdk[fastapi]==1.38.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
faker==20.1.0